set damage as double; add negative damage check;

+++ unit-2/homework/src/main/java/pro/it/sis/javacourse/DamageFactory.java
@@ -1,7 +1,7 @@
 package pro.it.sis.javacourse;
 
 public class DamageFactory {
-    public static Damage createDamage(DamageType type, int strength) {
+    public static Damage createDamage(DamageType type, double strength) {
         switch (type) {
             case physical:
                 return new PhysicalDamage(strength);
